tests-e2e/runner/requirements.txt
requests
pynacl
cbor2